feat: Section 18 — ERC-20 token transfer (USDC/USDT ABI encoding, EvmRpcAdapter asset branch)
- Add Erc20TransferEncoder: static ABI encoder for ERC-20 transfer(address,uint256)
  selector 0xa9059cbb, 32-byte left-zero-padded address, 32-byte big-endian uint256
- EvmRpcAdapter.prepareSend() branches on asset: native ETH vs ERC-20 token
  ERC-20 path: to=contractAddress, value=0, data=ABI calldata, gasLimit from AssetRegistryService
  Falls back to ETH native when AssetRegistryService is absent
- Fix RawTransaction.createTransaction() arg order (data before fee params)
- Fix String.format %64s padding (0 flag invalid for %s, use space-then-replace)
- CreateWithdrawalRequest.amount changed from BigDecimal to BigInteger (smallest unit)
- WithdrawalService validates asset via AssetRegistryService before creating withdrawal
- Tests: Erc20TransferEncoderTest (ABI encoding correctness), EvmRpcAdapterErc20Test
  (ETH native / USDC ERC-20 / fallback paths), plus evm/ subpackage variant
- V8 migration already seeds USDC/USDT with mainnet contract addresses and gasLimit=65000

+++ custody/src/test/java/lab/custody/orchestration/TestFixtures.java
@@ -2,7 +2,7 @@ package lab.custody.orchestration;
 
 import lab.custody.orchestration.CreateWithdrawalRequest;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 /**
  * 9-4-3: 공통 테스트 픽스처 클래스.
@@ -42,7 +42,7 @@ public final class TestFixtures {
      */
     public static CreateWithdrawalRequest evmRequest(String toAddress, String amount) {
         return new CreateWithdrawalRequest(
-                "evm", ADDR_FROM, toAddress, "USDC", new BigDecimal(amount));
+                "evm", ADDR_FROM, toAddress, "USDC", new BigInteger(amount));
     }
 
     /**
@@ -57,7 +57,7 @@ public final class TestFixtures {
      */
     public static CreateWithdrawalRequest evmRequestFrom(String fromAddress, String toAddress, String amount) {
         return new CreateWithdrawalRequest(
-                "evm", fromAddress, toAddress, "USDC", new BigDecimal(amount));
+                "evm", fromAddress, toAddress, "USDC", new BigInteger(amount));
     }
 
     /**
@@ -65,7 +65,7 @@ public final class TestFixtures {
      */
     public static CreateWithdrawalRequest bftRequest(String toAddress, String amount) {
         return new CreateWithdrawalRequest(
-                "bft", ADDR_FROM, toAddress, "USDC", new BigDecimal(amount));
+                "bft", ADDR_FROM, toAddress, "USDC", new BigInteger(amount));
     }
 
     /**
